# Pull performance fixes from semgrep