--- a/libs/commons/Hashtbl_.mli
+++ b/libs/commons/Hashtbl_.mli
@@ -2,6 +2,9 @@
    Extension of the standard library module Hashtbl
 *)
 
+type hash = int
+
+val combine_hash : hash -> hash -> hash
 val hash_of_list : ('a * 'b) list -> ('a, 'b) Hashtbl.t
 val hash_to_list : ('a, 'b) Hashtbl.t -> ('a * 'b) list
 val hkeys : ('a, 'b) Hashtbl.t -> 'a list