Pulling perf improvements from semgrep. In particular:

- https://github.com/semgrep/semgrep/commit/721ee143d58f4bf811e0d6b397f30621cbbb0dcf
- https://github.com/semgrep/semgrep/commit/0192a486332be1b396df2ab5198453c343498b74

Results for opengrep:

| test repo | current main | 46fe0ddc5c9909801d3b3b92502b93a3e8f8faf8 | diff(s) | diff(%) |
|---|---|---|---|---|
| grafana | 2712.33 | 2645.42 | -66.91 | -2.47 |
| go | 602.15 | 536.79 | -65.36 | -10.85 |
| rust | 827.96 | 769.77 | -58.19 | -7.03 |
| pmd | 87.39 | 88.03 | 0.64 | 0.73 |
| jellyfin | 79.05 | 78.83 | -0.22 | -0.27 |
| lemur | 33.70 | 33.81 | 0.12 | 0.34 |
| ComfyUI-Custom-Scripts | 14.13 | 13.74 | -0.39 | -2.79 |

(avg of 3 runs)